requirements.txt
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==3.1.0
Click==7.0
colorama==0.4.1
cropresize2==0.1.9
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
Flask==1.0.2
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.1
jieba==0.39
Jinja2==2.10
joblib==0.13.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
MarkupSafe==1.1.0
matplotlib==3.1.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.2
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
py2exe==0.9.2.2
Pygments==2.4.2
PyMySQL==0.9.3
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
python-magic==0.4.10
pytz==2019.2
pywinpty==0.5.5
pyzmq==18.1.0
qtconsole==4.5.5
scikit-learn==0.21.3
scipy==1.3.1
Send2Trash==1.5.0
short-url==1.2.1
six==1.12.0
sklearn==0.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.5.1
wordcloud==1.5.0
wechatpy==1.8.3
requests==2.22.0
flask_script==2.0.6
flask_migrate==2.5.2
cryptography==2.8